# docs/swagger.yaml
definitions:
...
  httputil.Error:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  user.LoginResponse:
    properties:
      expires_in:
        type: integer
      session_token:
        type: string
    type: object
  user.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  user.RegisterResponse:
    properties:
      email:
        type: string
      user_id:
        type: string
      verification_sent:
        type: boolean
    type: object
  user.RequestOTPRequest:
    properties:
      email:
        type: string
      phone:
        type: string
      purpose:
        description: '"login", "verify", "reset"'
        type: string
    type: object
  user.RequestOTPResponse:
    properties:
      expires_in:
        type: integer
      message:
        type: string
    type: object
  user.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        type: string
      phone:
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      enabled:
        type: boolean
      name:
        type: string
      phone:
        type: string
    type: object
  user.UserExportResponse:
    properties:
      consent_granted:
        type: boolean
      consent_timestamp:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      phone_verified:
        type: boolean
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  user.UserInfoResponse:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      phone_number_verified:
        type: boolean
      sub:
        type: string
    type: object
  user.UserSummary:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      enabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  user.VerifyOTPRequest:
    properties:
      code:
        type: string
      email:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password. Returns a session token.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Login
      tags:
      - auth
  /logout:
    post:
      description: Invalidate the current session token.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - SessionAuth: []
      summary: Logout
      tags:
      - auth
...
  /otp/request:
    post:
      consumes:
      - application/json
      description: Send a one-time password to the user's email or phone.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: OTP request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RequestOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.RequestOTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Request OTP
      tags:
      - auth
  /otp/verify:
    post:
      consumes:
      - application/json
      description: Verify the one-time password sent by /otp/request.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: OTP verification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Verify OTP
      tags:
      - auth
  /password/reset:
    post:
      consumes:
      - application/json
      description: Reset a user's password using an OTP code. Call /otp/request with
        purpose=reset first.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Password reset request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Reset password
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account within the resolved tenant.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.Error'
      summary: Register user
      tags:
      - auth
...
  /tenants/{tid}/users:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserSummary'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: List users
      tags:
      - users
  /tenants/{tid}/users/{uid}:
    put:
      consumes:
      - application/json
      description: Update a user's profile fields (name, phone, enabled status).
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: User ID
        in: path
        name: uid
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Update user
      tags:
      - users
  /tenants/{tid}/users/{uid}/export:
    get:
      description: Download all personal data held for a user as JSON (GDPR Art. 15/20).
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: User ID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserExportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Export user data (GDPR)
      tags:
      - users
...
  /tenants/{tid}/users/{uid}/purge:
    delete:
      description: Permanently delete all personal data for a user. Irreversible (GDPR
        Art. 17).
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: User ID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Purge user (GDPR)
      tags:
      - users
...
  /userinfo:
    get:
      description: Returns claims about the authenticated user (OIDC Core §5.3). Pass
        the session token as a Bearer token.
      parameters:
      - description: Tenant identifier
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - SessionAuth: []
      summary: UserInfo
      tags:
      - auth
swagger: "2.0"